Detect current network and netmask
Previously the network was guessed (assumed it was a /24 by replacing the last octet with a 0) and then later the netmask was hardcoded to 255.255.240.0 as a result the requested route is almost always incorrect. As most people supply their own or use the vpn tunnel as a default gateway it's probably rare to cause any issues.